SVM: Move transaction processing code out of bank.rs
