+++ runtime/src/svm/mod.rs
@@ -1,3 +1,4 @@
 pub mod account_loader;
 pub mod account_rent_state;
 pub mod transaction_account_state_info;
+pub mod transaction_processor;
